+++ autorest/codegen/models/code_model.py
@@ -409,3 +409,18 @@ class CodeModel:  # pylint: disable=too-many-instance-attributes, too-many-publi
                     request_builder.name = request_builder.name + "_initial"
                 operation.request_builder = request_builder
                 operation.link_body_kwargs_to_body_params()
+
+    def get_models_filename(self, is_python3_file: bool) -> str:
+        if (
+            self.options["version_tolerant"] or self.options["low_level_client"]
+        ) and self.options["python3_only"]:
+            return "_models"
+        if is_python3_file:
+            return "_models_py3"
+        return "_models"
+
+    @property
+    def enums_filename(self) -> str:
+        if self.options["version_tolerant"] or self.options["low_level_client"]:
+            return "_enums"
+        return f"_{self.module_name}_enums"